♻ 채팅방 이동할 때 서버 통신으로 딜레이되는 현상 수정
